chore: Update npm dependencies and add date-fns package
The code changes include updating the npm dependencies to their latest versions, such as "@trpc/server" and "clsx". Additionally, the "date-fns" package is added as a new dependency with version "^3.6.0". This chore ensures that the project is using the most up-to-date and stable versions of these dependencies and adds support for date manipulation and formatting with the "date-fns" library.