chore: Update npm dependencies and add react-loading-skeleton
The code changes include updating the npm dependencies to their latest versions, such as "@tanstack/react-query", "@trpc/client", "@trpc/next", "@trpc/react-query", "@trpc/server", and "zod". Additionally, the "react-loading-skeleton" package is added as a new dependency with version "^3.4.0". This chore ensures that the project is using the most up-to-date and stable versions of these dependencies and adds support for the "react-loading-skeleton" component.